% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_secret.R
\name{get_secret}
\alias{get_secret}
\title{get_secret
Retrieve a secret from the environment or using `op`}
\usage{
get_secret(secret_name, ...)
}
\arguments{
\item{secret_name}{character name of the key to retrieve}

\item{...}{additional arguments passed to `op`}
}
\value{
character, the requested secret
}
\description{
get_secret
Retrieve a secret from the environment or using `op`
}
\details{
This function retrieves a secret with the name specified by `secret_name`.
If `secret_name` is not an environmental variable, the function will attempt to retrieve the secret from from 1Password using `op`.

To temporarily set the secret as an environmental variable (i.e. for the duration of R the session) use `Sys.setenv(secret_name = "secret")`, where "secret" is the API key, and `secret_name` is the name of the secret (e.g. "GEMINI_API_KEY").
To permanently set the secret as an environmental variable, add `secret_name=secret` (no quotes) to your `.Renviron` file - it is recommended you use `usethis::edit_r_envrion()` to open the .Renviron file if you are setting this globally.
}
\seealso{
`?op`
}
